backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
dashscope==1.20.1
openai==1.45.0
pymupdf==1.24.9
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.4.0
aiofiles==24.1.0
httpx==0.27.2
alembic==1.13.3
greenlet==3.1.0
